feat: restore e621 command, add filters to e621 command
